initial commit of tremolo+bandpass prototype
initial commit of code. based on a multi effect sample code. blue led mode is the bandpass + tremolo with center f and q on the pots, green led mode is the svf with bandpass, center f and q on the pots. pink led mode is just the tremolo effect with depth and rate on the pots.